project/samd-userrow: reintroduce Makefile for reprogramming SAMD NVM User Row
Add a dedicated directory for producing an ELF image suitable
for re-programming the SAMD NVM User Row.

(Just manually tested this on a new Adafruit ItsyBitsy M0.)

+++ project/samd-userrow/userrow.c
@@ -1,5 +1,3 @@
-#include <bits.h>
-
 /* this file is used to generate
  * an SREC file to program the NVM User Row */
 
@@ -10,5 +8,5 @@
 #define USER_ROW_DEFAULT 0xFFFFFC5DD8E0C7FF
 
 __attribute__((section(".userrow")))
-u64 userrow = USER_ROW_DEFAULT;
+unsigned long long userrow = USER_ROW_DEFAULT;
